pyproject.toml
[tool.poetry]
name = "jesafe"
version = "0.1.0"
description = ""
authors = ["lwj0831 <wjl0831@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.121.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pandas = "^2.3.3"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
jupyterlab = "^4.4.10"
ipykernel = "^7.1.0"
httpx = "^0.28.1"
openpyxl = "^3.1.5"
rank-bm25 = "^0.2.2"
ddgs = "^9.9.1"
llama-parse = "^0.6.81"
llama-index-core = "^0.14.8"
pyjwt = "^2.10.1"
kiwipiepy = "^0.22.0"
requests = "^2.32.5"
nltk = "^3.9.2"
ragas = "^0.3.9"
datasets = "^4.4.1"
pyyaml = "^6.0.3"
deepl = "^1.25.0"
rapidfuzz = "^3.14.3"
alembic = "^1.17.2"
langchain = "^1.0.8"
langchain-openai = "^1.0.3"
langchain-community = "^0.4.1"
langchain-postgres = "^0.0.16"
cohere = "^5.20.0"
langchain-cohere = "^0.5.0"
langchain-classic = "^1.0.0"
unstructured = "^0.18.20"
markdown = "^3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
